# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Lifestyle Super App API
  description: |
    ## 🚀 Lifestyle Super App - REST API Documentation
    
    **Base URL:** `https://api.lifestyle.vn/v1`  
    **Local Dev:** `http://localhost:3000/api/v1`
    
    ### Authentication
    Hầu hết endpoints yêu cầu JWT Bearer token:
    ```
    Authorization: Bearer <access_token>
    ```
    
    ### Response Format
    ```json
    {
      "success": true,
      "data": { ... },
      "message": "Success",
      "meta": { "page": 1, "limit": 20, "total": 100 }
    }
    ```
    
    ### Error Format
    ```json
    {
      "success": false,
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Email is required",
        "statusCode": 400
      }
    }
    ```
    
    ### Modules
    | Module | Prefix | Description |
    |--------|--------|-------------|
    | Auth | `/auth` | Đăng ký, Đăng nhập, Token |
    | Users | `/users` | Hồ sơ, Địa chỉ, Preferences |
    | Drivers | `/drivers` | Đăng ký tài xế, Xe, Giấy tờ |
    | Merchants | `/merchants` | Cửa hàng, Nhân viên |
    | Wallets | `/wallets` | Ví, Nạp tiền, Rút tiền |
    | Payments | `/payments` | Thanh toán, Webhooks |
    | Spotlight | `/spotlight` | Redcomments, Reviews, Creators |
    | Loyalty | `/loyalty` | Xu, Referral |

  version: 1.0.0
  contact:
    name: Lifestyle Super App Team
    email: dev@lifestyle.vn
  license:
    name: Private
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development
...
security:
  - BearerAuth: []

# =============================================================================
# PATHS (API Endpoints)
# =============================================================================
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Đăng ký tài khoản mới
      description: |
        Tạo tài khoản mới bằng email + password hoặc số điện thoại.
        Tự động tạo Wallet cho user.
        Gửi OTP xác thực qua email/SMS.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              email_register:
                summary: Đăng ký bằng email
                value:
                  email: "nguyenvana@gmail.com"
                  password: "SecurePass@123"
                  firstName: "Nguyễn Văn"
                  lastName: "A"
                  phoneNumber: "0901234567"
              phone_register:
                summary: Đăng ký bằng SĐT
                value:
                  phoneNumber: "0901234567"
                  password: "SecurePass@123"
                  firstName: "Nguyễn Văn"
                  lastName: "A"
      responses:
        '201':
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email/SĐT đã tồn tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "USER_ALREADY_EXISTS"
                  message: "Email đã được sử dụng"
                  statusCode: 409
  /auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập
      description: Đăng nhập bằng email/SĐT + password. Trả về access token và refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              email_login:
                summary: Đăng nhập bằng email
                value:
                  email: "nguyenvana@gmail.com"
                  password: "SecurePass@123"
              phone_login:
                summary: Đăng nhập bằng SĐT
                value:
                  phoneNumber: "0901234567"
                  password: "SecurePass@123"
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Sai mật khẩu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "INVALID_CREDENTIALS"
                  message: "Email hoặc mật khẩu không đúng"
                  statusCode: 401
...
  /drivers/register:
    post:
      tags: [Drivers]
      summary: Đăng ký làm tài xế
      description: |
        User hiện tại đăng ký trở thành tài xế.
        Khởi tạo hồ sơ tài xế với trạng thái PENDING_VERIFICATION.
        Bắt đầu quy trình onboarding 8 bước.
      responses:
        '201':
          description: Đăng ký thành công, bắt đầu onboarding
          content:
            application/json:
              example:
                success: true
                data:
                  driverId: "uuid-driver-1"
                  driverNumber: "LS-DRV-00002"
                  status: "PENDING_VERIFICATION"
                  onboardingStep: 1
                  message: "Chào mừng! Hãy hoàn thành hồ sơ để bắt đầu nhận chuyến."
        '409':
          description: Đã đăng ký tài xế rồi
...
  /admin/drivers/{driverId}/verify:
    parameters:
      - name: driverId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Admin]
      summary: "[Admin] Duyệt/Từ chối tài xế"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [APPROVE, REJECT]
                rejectionReason:
                  type: string
            example:
              action: "APPROVE"
      responses:
        '200':
          description: Đã xử lý
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token. Lấy từ `/auth/login` hoặc `/auth/register`.
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIs...
        ```
    SepayWebhookKey:
      type: apiKey
      in: header
      name: X-Sepay-Key
      description: API Key xác thực webhook từ SePay

  # ---------------------------------------------------------------------------
  # Parameters (Reusable)
  # ---------------------------------------------------------------------------
...
  responses:
...
    ValidationError:
      description: Dữ liệu không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Dữ liệu không hợp lệ"
              statusCode: 400
              details:
                email: "Email không đúng định dạng"
                password: "Mật khẩu phải có ít nhất 8 ký tự"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            statusCode:
              type: integer
            details:
              type: object

    # --- Auth ---
    RegisterRequest:
      type: object
      required: [password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
          example: "nguyenvana@gmail.com"
        phoneNumber:
          type: string
          pattern: '^0[0-9]{9}$'
          example: "0901234567"
        password:
          type: string
          minLength: 8
          description: "Tối thiểu 8 ký tự, có chữ hoa + số + ký tự đặc biệt"
        firstName:
          type: string
          example: "Nguyễn Văn"
        lastName:
          type: string
          example: "A"
        referralCode:
          type: string
          description: Mã giới thiệu (optional)
          example: "LS8068F741"
    LoginRequest:
      type: object
      required: [password]
      properties:
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        password:
          type: string
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/UserSummary'
            accessToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIs..."
            refreshToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIs..."
            expiresIn:
              type: integer
              description: Access token TTL (seconds)
              example: 3600

    # --- User ---
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        phoneNumber:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [CUSTOMER, DRIVER, RESTAURANT_OWNER, SELLER, ADMIN]
        ekycLevel:
          type: string
          enum: [LEVEL_0, LEVEL_1, LEVEL_2, LEVEL_3]
        isActive:
          type: boolean
...